// enums/Games/GamesParameters.enum.ts
// Enum for platforms
export enum PlatformEnum {
  PC = "pc",
  BROWSER = "browser",
  ALL = "all",
}

// Enum for categories
export enum CategoryEnum {
  MMORPG = "mmorpg",
  SHOOTER = "shooter",
  STRATEGY = "strategy",
  ACTION = "action",
  RACING = "racing",
  SPORTS = "sports",
  MMO = "mmo",
  SURVIVAL = "survival",
  SOCIAL = "social",
}

// Enum for sort options
export enum SortByEnum {
  RELEASE_DATE = "release-date",
  POPULARITY = "popularity",
  ALPHABETICAL = "alphabetical",
  RELEVANCE = "relevance",
}